fix: messages 500, MinIO SigV4 PUT, Redis c-ares DNS, smoke test
- MessagesController: cast limit to long long — Drogon PG driver sent
  4-byte int for $2 causing "insufficient data left in message" from PG
- FilesController: replace broken AWS Signature V2 PUT auth with proper
  SigV4 Authorization header (signed headers: host, content-type,
  x-amz-date, x-amz-content-sha256); refactor signing key into helper
- main.cpp: pre-resolve Redis hostname with getaddrinfo() before passing
  to Drogon, bypassing c-ares which fails to resolve Docker hostnames
- infra/scripts: add smoke_test.py and diagnose.sh

+++ backend/src/controllers/MessagesController.cpp
@@ -105,9 +105,11 @@ void MessagesController::listMessages(const drogon::HttpRequestPtr& req,
     requireMember(chatId, me, [=, cb = std::move(cb)](bool isMember) mutable {
         if (!isMember) return cb(jsonErr("Not a member of this chat", drogon::k403Forbidden));
 
-        int limit = std::min(std::max(std::stoi(req->getParameter("limit").empty()
-                                                ? "50"
-                                                : req->getParameter("limit")), 1), 100);
+        // Use long long — Drogon's PG driver serialises int as 4 bytes but the
+        // wire-protocol parameter slot expects 8 bytes, causing "insufficient data".
+        long long limit = std::min(std::max(std::stoi(req->getParameter("limit").empty()
+                                                      ? "50"
+                                                      : req->getParameter("limit")), 1), 100);
         std::string beforeStr = req->getParameter("before");
 
         auto db = drogon::app().getDbClient();
